requirements-dev.txt
-r requirements.txt
google-generativeai==0.3.2
openai==1.12.0
pytest==8.0.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
